[1110] 抽离 liii-libaesgm/liii-pdfhummus 包定义
照 3rdparty/doctest.lua 的模式，把内联在根 xmake.lua 的两个
package 定义迁到 3rdparty/{libaesgm,pdfhummus}.lua，set_sourcedir
路径相应调整为相对 3rdparty/。顺手删除冗余的
PDFHUMMUS_VERSION（已在 xmake/vars.lua 定义）。
